Check the compatibility with a list (static)

+++ setup_ulp.py
@@ -32,7 +32,8 @@ lin_dir = "~/Library/Arduino15/packages/esp32"
 mac_dir = "~/Library/Arduino15/packages/esp32"
 
 wdir = "" # Working directory
-core_version = "1.0.0" # 1.0.0 by default, will be determinated after the OS
+# core_version = "1.0.0" # 1.0.0 by default, will be determinated after the OS
+supported_core_version = ['1.0.0', '1.0.1'] # We can imagine to update this variable from the directory
 
 """
 In the 'arduino_ulp' release you downloaded and unpacked, copy the folder 'ulp' to .../esp32/hardware/esp32/1.0.0/tools/sdk/include/ replacing the existing folder named 'ulp', "1.0.0" is the version number of the core you installed, change version number accordingly.
@@ -68,41 +69,57 @@ def main(argv):
 	
     else:
         print("Unsupported operating system: " + sys.platform)
-        exit(0)
+        exit(1)
 
-    ## Uncompress all the files
-    print("\nUncompress the files in temp")
+    ## Extract all the files
+    print("\nExtract the files in temp")
     unzip()
 
-    ## Find and print the core version
-    core_version = os.listdir(os.path.join(wdir,'hardware','esp32'))[0]
-    print("ESP32 core version: " + core_version)
+    ## Find and print the core version(s) and check compatibility
+    list_core_version = [en for en in os.listdir(os.path.join(wdir,'hardware','esp32')) if os.path.isdir(os.path.join(wdir,'hardware','esp32', en)) and en in supported_core_version]
+    list_len = len(list_core_version)
+    if list_len > 1:
+        print("ULP will be installed for these core versions of the esp32: V" + " - V".join(list_core_version))
+    elif list_len == 1:
+        print("ULP will be installed for this core version of the esp32: V" + " - V".join(list_core_version))
+    else:
+        print("No supported version detected. Install ESP32 Arduino core or check the check the Github page of the project for your issue.")
     
-    ## Copy the 'ulp' folder
-    print("\n(1/3) Copy the 'ulp' folder")
-    os.chdir(os.path.abspath('arduino_ulp-master'))
-    ulp_dir = os.path.join(wdir,'hardware','esp32', core_version, 'tools', 'sdk', 'include', 'ulp')
-    try:
-        shutil.rmtree(ulp_dir)
-    except:
-        print("No 'ulp' directory, 'ulp' will be created")
-    shutil.copytree(os.path.abspath('ulp'), ulp_dir)
-	
-    ## Copy the 'platform.txt' file
-    print("(2/3) Copy the 'platform.txt' file")
-    plt_dir = os.path.join(wdir,'hardware','esp32', core_version)
-    shutil.copyfile('platform.txt', os.path.join(plt_dir, 'platform.txt'))
 
-    ## Copy the 'binutils'
-    print("(3/3) Copy 'binutils' folder\n")
-    os.chdir(directory), os.chdir(temp_dir)
-    bin_dir = os.path.join(wdir,'tools', 'binutils', 'esp32ulp-elf-binutils')
-    try:
-        shutil.copytree(os.path.abspath('esp32ulp-elf-binutils'), bin_dir)
-    except:
-        print("Overwritte 'esp32ulp-elf-binutils'")
-        shutil.rmtree(bin_dir)
-        shutil.copytree(os.path.abspath('esp32ulp-elf-binutils'), bin_dir)
+    if list_len:
+        i = 1
+        for core_version in list_core_version:
+            ## Print core version
+            print("\nFor core version: " + core_version)
+            ## Copy the 'ulp' folder
+            print("(" + str(i) + "/" + str(list_len*2+1) + ") Copy the 'ulp' folder")
+            i += 1
+            os.chdir(os.path.join(directory, temp_dir ,'arduino_ulp-master'))
+            ulp_dir = os.path.join(wdir,'hardware','esp32', core_version, 'tools', 'sdk', 'include', 'ulp')
+            try:
+                shutil.rmtree(ulp_dir)
+            except:
+                print("No 'ulp' directory, 'ulp' will be created")
+            shutil.copytree(os.path.abspath('ulp'), ulp_dir)
+                
+            ## Copy the 'platform.txt' file
+            print("(" + str(i) + "/" + str(list_len*2+1) + ") Copy the 'platform.txt' file")
+            i += 1
+            plt_dir = os.path.join(wdir,'hardware','esp32', core_version)
+            shutil.copyfile('platform.txt', os.path.join(plt_dir, 'platform.txt'))
+
+        ## Copy the 'binutils'
+        print("\n(" + str(i) + "/" + str(list_len*2+1) + ") Copy 'binutils' folder\n")
+        os.chdir(directory), os.chdir(temp_dir)
+        bin_dir = os.path.join(wdir,'tools', 'binutils', 'esp32ulp-elf-binutils')
+        try:
+            shutil.copytree(os.path.abspath('esp32ulp-elf-binutils'), bin_dir)
+        except:
+            print("Overwrite 'esp32ulp-elf-binutils'")
+            shutil.rmtree(bin_dir)
+            shutil.copytree(os.path.abspath('esp32ulp-elf-binutils'), bin_dir)
+    else:
+        print("Installation failed")
 
     ## Remove the 'temp' directory
     print("Go back in the 'setup' directory and remove the 'temp' directory (" + temp_dir + ")")
@@ -110,6 +127,7 @@ def main(argv):
     shutil.rmtree(os.path.abspath(temp_dir))
     print("Execution time: " + str(time.time()-start_time) + " seconds")
     pause()
+    exit(0)
 
 ##  Function to create a directory
 ##  (https://gist.github.com/keithweaver/562d3caa8650eefe7f84fa074e9ca949)
